# swagger.yaml
swagger: "2.0"
info:
  description: |
    Backend API to support HotelHopper.cf \
    To see the standardized api responses see the models at the bottom\: 
    SuccessfulResponse for 200, BadRequestResponse for 400, and UnauthorizedResponse for 401 \
    More HTTP Status Codes may be added in the future and similar models will be made avialable.
  version: "1.0.0"
  title: "Hotel Hopper"
host: "staging-api.hotelhopper.cf"
...
schemes:
- "https"
paths:
  /auth/login:
    post:
      tags:
      - "auth"
      summary: "Returns auth token and user details"
      description: "Returns a new JWT token with a 24 hour expiry time. Use this token with all authentication requests using the Header('Authorization, 'bearer \\<JWT>')."
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
      - in: body
        name: body
        schema:
            type: object
            required:
              - email
              - password
            properties:
              email:
                type: string
              password: 
                type: string
      responses:
        200:
          description: "OK"
          schema:
            type: object
            example:
              error: false
              data:
                token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImNoYXJsZXNAc2hlYXJzaGFyZS5jb20iLCJpYXQiOjE1NTQ3OTAwNDd9.uLS_-XvoKyAkKHLxlWOA4WkVWf_j3oe_Ss_X2c_H6s4"
                user:
                  firstName: John
                  lastName: Doe
                  email: john.doe@email.com
                  isAdmin: false
        400:
          description: "Bad request"
          schema:
            $ref: '#/definitions/BadRequestResponse'
  /auth/register:
    post:
      tags:
      - "auth"
      summary: "Returns an auth token for further authentication"
      description: "Creates a new user and returns a JWT token with a 24 hour expiry time. Use this token with all authentication requests using the Header('Authorization, 'bearer \\<JWT>')."
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
      - in: body
        name: body
        schema:
            type: object
            required:
              - firstName
              - lastName
              - email
              - password
            properties:
              firstName:
                type: string
              lastName:
                type: string
              email:
                type: string
              password: 
                type: string
      responses:
        200:
          description: "OK"
          schema:
            type: object
            example:
              error: false
              data:
                token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImNoYXJsZXNAc2hlYXJzaGFyZS5jb20iLCJpYXQiOjE1NTQ3OTAwNDd9.uLS_-XvoKyAkKHLxlWOA4WkVWf_j3oe_Ss_X2c_H6s4"
        400:
          description: "Bad request"
          schema:
            $ref: '#/definitions/BadRequestResponse'
  /auth/userDetails:
    get:
      tags:
      - "auth"
      summary: "Returns user details based on JWT token"
      description: "Returns user details based on JWT token. The JWT token must not be expired."
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
      - name: Authorization
        in: header
        description: JWT bearer token given on login or signup. Should be in the format 'bearer \<JWT>'.
        required: true
        type: string
      responses:
        200:
          description: "OK"
          schema:
            type: object
            example:
              error: false
              data:
                user:
                  userId: 1
                  firstName: John
                  lastName: Doe
                  email: John.Doe@email.com
                  isAdmin: false
                  rewardPoints: 580
        401:
          description: "Unauthorized"
          schema:
            $ref: '#/definitions/UnauthorizedResponse'
  /auth/forgotPassword:
    post:
      tags:
      - "auth"
      summary: "Sends a reset password link to the provided email."
      description: "Sends a reset password link to the provided email."
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
      - in: body
        name: body
        schema:
            type: object
            required:
              - email
            properties:
              email:
                type: string
      responses:
        200:
          description: "OK"
          schema:
            type: object
            example:
              error: false
              message: "We've sent a link to your email to reset your password. It expires in 24 hours."
        400:
          description: "Bad Request"
          schema:
            $ref: '#/definitions/BadRequestResponse'
  /auth/resetPassword/<token>:
    get:
      tags:
      - "auth"
      summary: "Verifies the password reset token is still valid"
      description: "Verifies the password reset token is still valid"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      responses:
        200:
          description: "OK"
          schema:
            type: object
            example:
              error: false
              data:
                token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImNoYXJsZXNAc2hlYXJzaGFyZS5jb20iLCJpYXQiOjE1NTQ3OTAwNDd9.uLS_-XvoKyAkKHLxlWOA4WkVWf_j3oe_Ss_X2c_H6s4"
        400:
          description: "Bad Request"
          schema:
            $ref: '#/definitions/BadRequestResponse'
  /auth/resetPassword:
    patch:
      tags:
      - "auth"
      summary: "Resets the users password given a valid reset token"
      description: "Resets the users password given a valid reset token."
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
      - in: body
        name: body
        schema:
            type: object
            required:
              - password
              - confirmPassword
              - token
            properties:
              password:
                type: string
              confirmPassword:
                type: string
              token:
                type: string
      responses:
        200:
          description: "OK"
          schema:
            type: object
            example:
              error: false
              message: "Your password has been succesfully updated. Please log in to get a new token."
        400:
          description: "Bad Request"
          schema:
            $ref: '#/definitions/BadRequestResponse'
...
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
...
  BadRequestResponse:
    type: object
    properties:
      error:
        type: boolean
      message:
        type: string
        description: "The first found validation error"
      validationErrors:
        type: object
        properties:
          <name of parameter>:
            type: array
            description: "Array of the problems with the parameter value"
            items:
              type: string
    required:
    - error
    - message
  UnauthorizedResponse:
    type: object
    properties:
      error:
        type: boolean
      message:
        type: string
        enum:
          - "No user found with JWT"
          - "Invalid JWT format"
          - "Expired JWT"
    required:
    - error
    - message
...